Fix: Refactor codes for Search.js file To improve the structure and apply Styled-components appropriately Ref: #134
