+++ src/components/ScriptItem/style.js
