+++ src/shared/App.js
@@ -15,7 +15,7 @@ import MyAllDictList from '../components/MyDict/MyAllDictList';
 import GoogleAnalytics from '../components/GoogleAnalytics';
 import KakaoRedirect from '../pages/Kakaoredirect';
 import ScriptFiltering from '../pages/ScriptFiltering';
-import Search from '../pages/Search';
+import Search from '../pages/Search/Search';
 import NotFound from '../pages/NotFound';
 import Main from '../pages/Main';
 import Typing from './../pages/Typing/Typing';
